\name{myLDA-package}
\alias{myLDA-package}
\alias{myLDA}
\docType{package}
\title{
Implementation of LDA methods for my Master's Thesis
myLDA 
}
\description{
More about what it does (maybe more than one line)
A concise (1-5 lines) description of the package 
}
\details{
\tabular{ll}{
Package: \tab myLDA2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-01\cr
License: \tab What license is it under?\cr
}

}
\author{
Philipp Burckhardt

Maintainer: Who to complain to <yourfault@somewhere.net>
The author and/or maintainer of the package 
}
\references{
Literature or other references for background information 
}
Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory 
\keyword{ package }
\seealso{
Optional links to other man pages, e.g. 
\code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
simple examples of the most important functions 
}
